Move runtime/vectorized to runtime/
