+++ be/src/formats/csv/converter.h
@@ -15,7 +15,7 @@
 #include "runtime/date_value.hpp"
 #include "runtime/decimalv2_value.h"
 #include "runtime/timestamp_value.h"
-#include "runtime/vectorized/time_types.h"
+#include "runtime/time_types.h"
 #include "util/raw_container.h"
 #include "util/slice.h"
