+++ be/src/runtime/chunk_cursor.cpp
@@ -1,6 +1,6 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
 
-#include "chunk_cursor.h"
+#include "runtime/chunk_cursor.h"
 
 #include <utility>
