+++ be/src/storage/types.cpp
@@ -27,7 +27,7 @@
 #include "runtime/datetime_value.h"
 #include "runtime/decimalv2_value.h"
 #include "runtime/mem_pool.h"
-#include "runtime/vectorized/time_types.h"
+#include "runtime/time_types.h"
 #include "storage/collection.h"
 #include "storage/convert_helper.h"
 #include "storage/decimal12.h"
