+++ be/test/column/date_value_test.cpp
@@ -30,7 +30,7 @@
 #include "column/fixed_length_column.h"
 #include "runtime/date_value.h"
 #include "runtime/timestamp_value.h"
-#include "runtime/vectorized/time_types.h"
+#include "runtime/time_types.h"
 
 namespace starrocks {
 namespace vectorized {
