+++ be/test/exprs/vectorized/cast_expr_test.cpp
@@ -14,7 +14,7 @@
 #include "gen_cpp/Exprs_types.h"
 #include "gen_cpp/Types_types.h"
 #include "runtime/primitive_type.h"
-#include "runtime/vectorized/time_types.h"
+#include "runtime/time_types.h"
 #include "util/json.h"
 
 namespace starrocks {
