+++ be/test/exprs/vectorized/decimal_cast_expr_test_helper.h
@@ -12,7 +12,7 @@
 #include "exprs/vectorized/cast_expr.h"
 #include "exprs/vectorized/mock_vectorized_expr.h"
 #include "runtime/primitive_type.h"
-#include "runtime/vectorized/time_types.h"
+#include "runtime/time_types.h"
 #include "testutil/parallel_test.h"
 
 namespace starrocks::vectorized {
