+++ be/test/exprs/vectorized/es_functions_test.cpp
@@ -9,7 +9,7 @@
 #include "column/fixed_length_column.h"
 #include "exprs/vectorized/mock_vectorized_expr.h"
 #include "runtime/runtime_state.h"
-#include "runtime/vectorized/time_types.h"
+#include "runtime/time_types.h"
 #include "testutil/function_utils.h"
 
 namespace starrocks {
