+++ be/test/exprs/vectorized/time_functions_test.cpp
@@ -15,7 +15,7 @@
 #include "column/vectorized_fwd.h"
 #include "exprs/vectorized/mock_vectorized_expr.h"
 #include "runtime/runtime_state.h"
-#include "runtime/vectorized/time_types.h"
+#include "runtime/time_types.h"
 #include "testutil/function_utils.h"
 #include "udf/udf.h"
