+++ be/test/runtime/vectorized/sorted_chunks_merger_test.cpp
@@ -1,6 +1,6 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
 
-#include "runtime/vectorized/sorted_chunks_merger.h"
+#include "runtime/sorted_chunks_merger.h"
 
 #include <gtest/gtest.h>
