+++ be/test/storage/task/engine_storage_migration_task_test.cpp
@@ -14,7 +14,7 @@
 #include "runtime/mem_tracker.h"
 #include "runtime/memory/chunk_allocator.h"
 #include "runtime/user_function_cache.h"
-#include "runtime/vectorized/time_types.h"
+#include "runtime/time_types.h"
 #include "storage/chunk_helper.h"
 #include "storage/options.h"
 #include "storage/rowset/rowset_factory.h"
